Add guards around creation/removal of snet hosts entry
Currently, the tasks in glance_snet_override are applied even when
glance_default_store is not set to 'swift'.  This minor change
updates the when condition so that we only add/remove the snet hosts
entry when glance_default_store == 'swift' AND when we have a key for
glance_swift_store_region in cf_snet_endpoints.